FIX: Infinite Loop caused by 16-bits overflow when code not found

+++ plugin/include/transform.h
@@ -8,4 +8,4 @@ uint16_t modified_to_utf16(uint16_t code);
 // FUN_08884680
 uint16_t sjis_to_utf16(uint16_t sjis);
 // FUN_08884724
-int binary_search(uint16_t sjis, uint16_t low, uint16_t high);
+int binary_search(uint16_t sjis, int low, int high);
